Acquire Input Data
Added functionality for taking input data. The user will be able to
enter a string of numbers, in the format [1,2,3] and it will parse it
into an array of just numbers. From there it will take the numbers and
group them into game frames (2 rolls). If there are 21 numbers in the
string it will put 3 rolls in the 10th game frame.